Move signal handler to concurrency
