+++ cmd/silkworm.cpp
@@ -1,5 +1,5 @@
 /*
-    Copyright 2021 The Silkworm Authors
+    Copyright 2021-2022 The Silkworm Authors
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@
 
 #include <silkworm/common/log.hpp>
 #include <silkworm/common/settings.hpp>
-#include <silkworm/common/signal_handler.hpp>
+#include <silkworm/concurrency/signal_handler.hpp>
 #include <silkworm/db/stages.hpp>
 #include <silkworm/stagedsync/sync_loop.hpp>
